coral-hf-spaces (1)/coral-deploy/space3-classification/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
Pillow>=9.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
python-multipart>=0.0.6